refactor: relocate spacy
